Project #5 Water Efficiency Calculator
The basic objective of the Water Efficiency Calculator project is to help people to find out how
much water is efficient to drink. Water Efficiency Calculator is a Windows-based application
that is designed to help users to determine how much the drinking water is effective and harmless
or on the other side, how much it is harmful and dangerous for health.